JSON Packets
The structure of the JSON packets has been amended. There is now a
"cdm" field that tells us what kind of packet it is. Available cmd
are:

• ping
• pong
• msg

Which is self-explanatory. When saving a packet in a Message instance
the cmd is saved, except for msg, where the message itself is saved.
I think it'll be better to amend the Message structure and save msg
anyway, and, if available, the msg, separately. TBD.

sendPing() and sendPong() have been rewritten, and simplified. A new
function, prepareJSONPacket(char*);, transforms a char array into a
JSON packet, to be sent to sendJSONPacket(). sendPing() and sendPong()
also use sendJSONPacket().