BINARY versions
I am adding precompiled binaries that can be uploaded with the folowwing command:

/Path/To/Arduino15/packages/electroniccats/tools/bossac/1.0.0-electroniccats-1/bossac -i -d --port=<PORT> -U true -i -e -w -v ./<FIRMWARE NAME>.bin -R

Slava is a pretty neutral version: no extras, and it logs to the Serial monitor.
Slavapas (a pun on Slava + ça va pas, it's not going well) is a version with HDC1080/CCS811 and an OLED, and either the code has an issue, or the hardware I am using right now is faulty. I suspect a hardware issue. I'll be doing some tests.

Pavel is my BastWAN for distance tests. Autoreply, auto PING every 2 minutes.